rename navigation file
