+++ bikeRentalMobile/App.tsx
@@ -3,7 +3,7 @@ import { SafeAreaProvider } from "react-native-safe-area-context";
 
 import useCachedResources from "./hooks/useCachedResources";
 import useColorScheme from "./hooks/useColorScheme";
-import Navigation from "./navigation";
+import Navigation from "./navigation/Navigation";
 
 export default function App() {
   const isLoadingComplete = useCachedResources();
